requirements.txt
openai>=1.40.0
pydantic>=2.8.2
python-dotenv>=1.0.1
slack-bolt>=1.19.0
slack-sdk>=3.27.1
requests>=2.32.3